crypto/message-digest: add common infrastructure, improve non-threaded.
this moves the common bits from linux-kcapi and openssl to a single
file, allows that cumbersome core to be shared with other
implementations.

it also introduces a maximum limit per feed if not using threads, this
way we avoid locking the main loop for too long.
